Problem: Skipped notifications to imported-template
An optimization in _notifyPath causes notifications not to be forwarded to imported-template if the value passed to _notifyPath is an object that is the same as the existing model. This is a legitimate case, in which dom-bind "resets" the object in order to inform that one of the properties has changed

Solution: always forward the notifications if the model is the same instance. Also fixed a side problem that appeared while fixing this. Starcounter-include forwarded notifications to imported-template before the template was stamped